Create navbar
This file just contains the page navigation links to be included at the top of each page. I put it in its own file so that if we ever want to add more pages, we just have to edit this one file instead of having to change all of the pages.