Update index.php
I know I had a reason for doing it the way I did when I first added this button, but I don't know what it was. As it is, the CSS transition isn't working correctly; this fixes that. I suppose it's due to the changes I've made to 'CSS/main.css'.