added support for BorderStyle 3
It may have some issues, but I'm not sure how to fix them. It also may not have any issues. I didn't have a subtitle file with BorderStyle 3 to test on, so I just changed the BorderStyle of another file. Some things did not display well, but the file wasn't made for BorderStyle 3, so that may be the real issue.