fixed \kf, mostly
SVG gradients are based on the length of the entire text element, not just a single tspan. So if there are multiple \kf's in one line, they won't be applied correctly. Also, if someone decided to use \kf with \t(some color), it won't work. Gradients can't be transitioned normally, so I'll have to write some code just for them.