Menu fixes and larger buttons
Apparently people still weren't seeing the controls even though we raised the opacity a while back. It opacity is now full and I made them a bit larger.

Also the menu looks a bit less shit now. A couple OCD fixes (Like the margins being uneven when considering the progressbar and the close menu button not aligning with the open menu button) - I also made the keybinds a bit less distracting while I was at it. People were complaining that the menu was messy, so this is an attempt to better emphasize important content.