fix url parsing/decoding, especially for embeds
The URL for an embed should be encoded correctly so we shouldn't have to handle that. It may have other GET parameters though, so we can't just treat the entire query string as an identifier. It will fall back to the previus method if it can't find a video from the first method.