use getComputedTextLength() instead of getBBox()
Also cache the height of the font when it's italic. I don't think this is any faster, and the memory savings from the function change are possibly offset by caching the italic font heights, but it feels like it *should* be better (and I've already written it), so I'm gonna keep it.