rework how split lines are stored
Previously all lines and line pieces used the same Subtitle class, and the Renderer had to keep track of which pieces went with which lines. Now the Subtitle class has been renamed LinePiece, and the new SubtitleLine class sits between it and the Renderer class. This commit is just enough to get things working as they were before; there are still a lot of improvements that can now be made due to this change.