remove parentheses from overrides
Transitions were the only overrides that actually needed parentheses, so I've a new override - \te - to mark the end of transitions. Transitions can also be ended by another transition or the end of the block. Since this can all be done at the start, it simplifies some of the parsing later on, and allows for font and style names to have parentheses in them.