commit video index generation code to document
This code doesn't work the way I wanted it to, so I removed it, but I'm going to commit it here before removing it again to document what happened.

The idea was that a video could be uniquely identified either by it's UID or by a seed and index pair. This turned out to not be possible for both theoretical and practical reasons, though I spent a long time working on it before realizing that. A video's UID *does* uniquely identify it, but a seed and index do not.

The reason it doesn't work theoretically is that when getting a random video, the seed and index are first used to determine which "behavior" the video will have, and then used again to determine which video with that behavior to return. Determining the video's behavior is percentage-based, but completely random. As such, it is entirely possible for a video to not be identified by *any* seed/index pair because that video's behavior could theoretically never be chosen.

There are a few reasons it doesn't work practically. One is that it is entirely possible to randomly choose a video that cannot normally be chosen. For example, a video with a behavior that is to be shown on an interval but otherwise never be randomly chosen can only be chosen if that behavior's "last seen" value is high enough. That is, choosing such a video depends on more than just the seed and index. Another reason it doesn't work practically is that, while it would be possible to determine a valid index for a chosen video most of the time, it could take a lot of resources to generate a good index. Like with the theoretical example above, if the correct behavior is never chosen the HTTP request could end up timing out first, all the while using processing resources.